refactor(server): move the open pipeline behind the ADR-034 seam
ADR-034 Unit 7a, the move the previous three commits were clearing a path
for. `documents/open.ts` stops forwarding to `mcp/file-opener.ts` and
becomes the pipeline: path resolution, session restore, content
finalization, and the disk/upload/scratchpad entries.

The fourth named entry, `openFromRestore`, is what closes the unit.
`restoreOpenDocuments` reached file-opener through a dynamic import whose
only job was dodging the cycle file-opener's own static import of
document-service created; with the pipeline below both, that import is
static and document-service now has zero dynamic imports. Naming the
restore path also gives the `readOnly` flag somewhere to live -- losing it
there is a bug that has shipped once (#1591).

Two things the guards caught that are worth reading:

The seam was exporting the implementations under BOTH their old names and
the aliases, so retargeting the tests would have carried the old
vocabulary across intact. The implementations now take the seam's names
outright. `SUPPORTED_EXTENSIONS` stopped being re-exported by a module
that does not own it, and `ensureAutoSave` got its own module rather than
riding a published surface -- it is the last thing in documents/ that
still reaches document-service, and leaving it on open.ts would have put
that edge straight back on the pipeline that just shed it.

Two controls in documents-open.test.ts failed on this correct change,
both because they were pinned to incidentals: a `> 3` count that was the
old export surface's size wearing a control's clothing, and a positive
control naming the one call site this commit removes. Both are now
derived from SANCTIONED itself, which also makes a zero-of-zero row --
the bug that derivation was introduced to fix -- impossible rather than
merely unlikely.
