+++ docs/decisions.md
@@ -639,7 +639,7 @@ Public registry interface (sketch — not authoritative):
 
 ## ADR-034: File-Open Pipeline with Named Entry Points and Shared Core
 
-**Status:** Accepted; partially implemented (verified against `src/` 2026-05-25). Part 1 landed: `src/server/documents/open.ts` exposes named entry aliases (`openFromDisk` / `openFromUpload` / `openScratchpad`) that forward to `file-opener.ts`, plus a derived `kindOfOpenResult` helper. **Unit 6 of the 2026-08-24 maintainability programme made the seam mandatory rather than opt-in**: every non-restore disk caller and every upload and scratchpad caller now imports it, `mcp/document.ts`'s duplicate re-export of the same three entries is deleted, and `tests/server/documents-open.test.ts` fails on any module under `src/` that reaches `mcp/file-opener.ts` outside a written-down exception list. Unit 6 also characterized every entry path first (`tests/server/adr-034-open-characterization.test.ts`), which is what makes the remaining parts verifiable as behaviour-preserving. **Deferred:** the shared internal pipeline still lives in the ~1000-line `file-opener.ts`, the `openFromRestore` entry point is not yet exposed, and `OpenResult` remains a derived enum rather than a discriminator on the result type. Pairs with ADR-033. Designed in the `/improve-codebase-architecture` grilling pass, 2026-05-15.
+**Status:** Accepted; largely implemented (verified against `src/` 2026-08-27). Part 1 landed: `src/server/documents/open.ts` exposes named entry aliases (`openFromDisk` / `openFromUpload` / `openScratchpad`) that forward to `file-opener.ts`, plus a derived `kindOfOpenResult` helper. **Unit 6 of the 2026-08-24 maintainability programme made the seam mandatory rather than opt-in**: every non-restore disk caller and every upload and scratchpad caller now imports it, `mcp/document.ts`'s duplicate re-export of the same three entries is deleted, and `tests/server/documents-open.test.ts` fails on any module under `src/` that reaches `mcp/file-opener.ts` outside a written-down exception list. Unit 6 also characterized every entry path first (`tests/server/adr-034-open-characterization.test.ts`), which is what makes the remaining parts verifiable as behaviour-preserving. **Unit 7a then moved the pipeline itself.** `documents/open.ts` now holds path resolution, session restore, content finalization and the three disk/upload/scratchpad entries, and exposes the fourth named entry `openFromRestore` — which is what let `document-service.ts` drop all three of its dynamic `await import("./file-opener.js")` calls, so the cycle they existed to dodge is gone rather than deferred. Four sibling modules came out of the same split: `documents/populate.ts` (content in and out of a Y.Doc), `documents/annotation-wiring.ts`, `documents/conflict.ts` (the external-conflict flag, whose read and write halves had been in different files) and `documents/watcher.ts` (the reload lifecycle, carrying the concurrent-reload guard as a named acquire/release contract). What is left in `mcp/file-opener.ts` is the reload family — three entries that replace an already-open document's content — and Unit 7c decides where those land. **Deferred:** `OpenResult` remains a derived enum rather than a discriminator on the result type (Unit 7b). Pairs with ADR-033. Designed in the `/improve-codebase-architecture` grilling pass, 2026-05-15.
 
 **Context:** `src/server/mcp/file-opener.ts` is 1049 lines exposing three public entry points (`openFileByPath`, `openFileFromContent`, `openScratchpad`) and internal helpers (`applyPreparedContent`, `clearAndReload`, `wireAnnotationStore`, `ensureAutoSave`). Six callers invoke `openFileByPath`: `startup-file.ts` (cold-start file-association), `index.ts` (welcome/changelog auto-open), `mcp/routes/open.ts` (HTTP REST API), `mcp/document.ts` (`tandem_open` MCP tool), `mcp/convert.ts` (after `.docx` HTML conversion), and `mcp/document-service.ts` (session restore — using a dynamic `await import(...)` to dodge a circular dependency through provider/registry state). Each caller wires the same downstream steps (track the doc, broadcast, set active, attach auto-save) in slightly different orders. `OpenFileResult` conflates outcomes via booleans (`forceReloaded`, `alreadyOpen`). The "open before HTTP bind" startup invariant (CLAUDE.md) is enforced by call ordering in `src/server/index.ts` only.
