+++ docs/gotchas.md
@@ -95,7 +95,7 @@ source: `y-prosemirror.cjs:834` merges adjacent Y.Text nodes in an untagged tran
 - **Authorship stamps COALESCE, so one entry is a run of typing rather than one keystroke — and three of its conditions fail silently if changed.** The map gained an entry per doc-changing transaction and was never compacted; anchored entries are never reaped (#1480), so nothing removed them. Measured: 500 distinct keys encode to 136,290 bytes of session file, 500 re-sets of one key to 310 — but **that 440x is a property of `doc.gc`**, and setting `gc: false` anywhere to preserve history would take it away with no other signal. `coalesceIntoPrevious` extends the previous entry when the author matches and the new text is adjacent, assembling the merged anchor from the two endpoint anchors that **already exist** rather than re-minting (re-deriving a known-good anchor crosses the PM-Y seam of #1450/#1459, and `flatOffsetToRelPos`'s assoc-directed retry returns a *wrong* anchor rather than none). Adjacency is compared in **PM positions**; the candidate is found by one id in extension storage, never a map scan, since an O(entries) fix would defeat its own purpose. The three silent ones: `lastStampId` is **left alone by a transaction that stamps nothing**, because backspace is exactly that and backspace-then-retype is the commonest editing rhythm there is — clearing it there guts the saving; it is **cleared before the y-sync early return**, or a remote/MCP/force-reload change leaves a candidate the next keystroke extends; and there is deliberately **no collapsed-candidate guard**, because measuring showed its only effect is to strand the collapsed entry forever, which is the accumulation being fixed.
 - **An annotation's `textSnapshot` may be a PREFIX, not the whole range — ask `isSnapshotTruncated()` before doing anything with it (#1486).** `captureSnapshot` caps it at 200 characters to bound record size against pathological ranges (#1000 review R2), and since #1486 reports the cut as `textSnapshotTruncated` rather than marking it in the text. **A prefix locates the range's START and says nothing about its END**, and that asymmetry decides what each consumer may do — the predicate and `snapshotSearchPrefix()` live in `src/shared/snapshot.ts` precisely so the four of them cannot drift apart:
   - **Undo** (`useAnnotationReview.svelte.ts`) needs the whole text and must **decline** — restoring a prefix deletes everything past the cap. The decline is permanent, so it reports via `onUndoFailed`, not just a console warning.
-  - **The reload relocation pass** (`file-opener.ts`) may **relocate on the prefix and carry the original span across**. Both simpler options are wrong: re-anchoring to `match + snapshot.length` shrinks a long annotation to 200 chars on every reload (after which accept replaces the wrong span), and *skipping* is no better, because `refreshAllRanges` has already minted a fresh `relRange` from the stale offsets — so the record ends up durably pinned to the wrong text and never revisited.
+  - **The reload relocation pass** (`documents/watcher.ts`) may **relocate on the prefix and carry the original span across**. Both simpler options are wrong: re-anchoring to `match + snapshot.length` shrinks a long annotation to 200 chars on every reload (after which accept replaces the wrong span), and *skipping* is no better, because `refreshAllRanges` has already minted a fresh `relRange` from the stale offsets — so the record ends up durably pinned to the wrong text and never revisited.
   - **The `.docx` apply guard** (`file-io/docx-apply.ts`) must compare with `startsWith`, not `===`. Exact comparison rejects every suggestion over a 200-char range and reports it as "the text changed" when nothing had.
   - **The chat anchor** (`App.svelte`) is deliberately outside all of this and keeps its in-band ellipsis — not because it stays local (it reaches the AI through `tandem_checkInbox`) but because nothing writes it back into the document. `SnapshotBearing` requires an `id` so it cannot be passed to these functions by mistake.
