+++ tests/server/file-opener-edge-cases.test.ts
@@ -3,16 +3,12 @@ import fs from "fs/promises";
 import os from "os";
 import path from "path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { openFromDisk, openFromUpload } from "../../src/server/documents/open.js";
 import { addDoc, removeDoc, setActiveDocId } from "../../src/server/documents/registry-testing.js";
 import { getOpenDocs } from "../../src/server/mcp/document-service.js";
-import {
-  openFileByPath,
-  openFileFromContent,
-  SUPPORTED_EXTENSIONS,
-} from "../../src/server/mcp/file-opener.js";
 import { handleOpen } from "../../src/server/mcp/routes/open.js";
 import { getOrCreateDocument } from "../../src/server/yjs/provider.js";
-import { Y_MAP_DOCUMENT_META } from "../../src/shared/constants.js";
+import { SUPPORTED_EXTENSIONS, Y_MAP_DOCUMENT_META } from "../../src/shared/constants.js";
 
 let tmpDir: string;
 
@@ -24,7 +20,7 @@ beforeEach(async () => {
   tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), "tandem-test-"));
 });
 
-describe("openFileByPath — file size limit", () => {
+describe("openFromDisk — file size limit", () => {
   it("rejects files exceeding 50MB", async () => {
     const bigFile = path.join(tmpDir, "big.md");
     const handle = await fs.open(bigFile, "w");
@@ -32,7 +28,7 @@ describe("openFileByPath — file size limit", () => {
     await handle.close();
 
     try {
-      await openFileByPath(bigFile);
+      await openFromDisk(bigFile);
       expect.unreachable("should have thrown");
     } catch (err: unknown) {
       const e = err as NodeJS.ErrnoException;
@@ -46,19 +42,19 @@ describe("openFileByPath — file size limit", () => {
     const smallFile = path.join(tmpDir, "small.md");
     await fs.writeFile(smallFile, "# Hello\nSmall content");
 
-    const result = await openFileByPath(smallFile);
+    const result = await openFromDisk(smallFile);
     expect(result.fileName).toBe("small.md");
     expect(result.format).toBe("md");
   });
 });
 
-describe("openFileByPath — unsupported extensions", () => {
+describe("openFromDisk — unsupported extensions", () => {
   it("rejects .csv with UNSUPPORTED_FORMAT code", async () => {
     const csvFile = path.join(tmpDir, "data.csv");
     await fs.writeFile(csvFile, "a,b,c\n1,2,3");
 
     try {
-      await openFileByPath(csvFile);
+      await openFromDisk(csvFile);
       expect.unreachable("should have thrown");
     } catch (err: unknown) {
       const e = err as NodeJS.ErrnoException;
@@ -70,28 +66,28 @@ describe("openFileByPath — unsupported extensions", () => {
   it("rejects .pdf files", async () => {
     const pdfFile = path.join(tmpDir, "doc.pdf");
     await fs.writeFile(pdfFile, "fake pdf");
-    await expect(openFileByPath(pdfFile)).rejects.toThrow("Unsupported file format");
+    await expect(openFromDisk(pdfFile)).rejects.toThrow("Unsupported file format");
   });
 
   it("rejects .js files", async () => {
     const jsFile = path.join(tmpDir, "code.js");
     await fs.writeFile(jsFile, "console.log('hello')");
-    await expect(openFileByPath(jsFile)).rejects.toThrow("Unsupported file format");
+    await expect(openFromDisk(jsFile)).rejects.toThrow("Unsupported file format");
   });
 });
 
-describe("openFileByPath — nonexistent files", () => {
+describe("openFromDisk — nonexistent files", () => {
   it("throws ENOENT for missing files", async () => {
     const missing = path.join(tmpDir, "ghost.md");
-    await expect(openFileByPath(missing)).rejects.toThrow();
+    await expect(openFromDisk(missing)).rejects.toThrow();
   });
 });
 
-describe("openFileByPath — supported extensions", () => {
+describe("openFromDisk — supported extensions", () => {
   it("opens .md files", async () => {
     const f = path.join(tmpDir, "test.md");
     await fs.writeFile(f, "# Hello");
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
     expect(result.format).toBe("md");
     expect(result.readOnly).toBe(false);
     expect(result.source).toBe("file");
@@ -100,7 +96,7 @@ describe("openFileByPath — supported extensions", () => {
   it("opens .txt files", async () => {
     const f = path.join(tmpDir, "test.txt");
     await fs.writeFile(f, "plain text");
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
     expect(result.format).toBe("txt");
     expect(result.readOnly).toBe(false);
   });
@@ -108,32 +104,32 @@ describe("openFileByPath — supported extensions", () => {
   it("opens .html files", async () => {
     const f = path.join(tmpDir, "test.html");
     await fs.writeFile(f, "<p>hello</p>");
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
     expect(result.format).toBe("html");
   });
 });
 
-describe("openFileByPath — duplicate detection", () => {
+describe("openFromDisk — duplicate detection", () => {
   it("returns alreadyOpen for the same file opened twice", async () => {
     const f = path.join(tmpDir, "dup.md");
     await fs.writeFile(f, "# Dup");
 
-    const first = await openFileByPath(f);
+    const first = await openFromDisk(f);
     expect(first.alreadyOpen).toBe(false);
 
-    const second = await openFileByPath(f);
+    const second = await openFromDisk(f);
     expect(second.alreadyOpen).toBe(true);
     expect(second.documentId).toBe(first.documentId);
   });
 
   it("reopening a renamed file by its NEW path resolves to the same tab, not a duplicate (#1017)", async () => {
     // After a rename the doc keeps its OLD documentId (= docIdFromPath(oldPath))
-    // but its registry filePath is the NEW path. A later openFileByPath(newPath)
+    // but its registry filePath is the NEW path. A later openFromDisk(newPath)
     // computes a DIFFERENT id and would open a duplicate tab without the realpath
     // fallback. Simulate the post-rename registry state, then reopen by new path.
     const oldPath = path.join(tmpDir, "before.md");
     await fs.writeFile(oldPath, "# Body");
-    const first = await openFileByPath(oldPath);
+    const first = await openFromDisk(oldPath);
     expect(first.alreadyOpen).toBe(false);
 
     // Rename on disk + reflect the promote-in-place registry state (same id).
@@ -143,7 +139,7 @@ describe("openFileByPath — duplicate detection", () => {
     if (!state) throw new Error("doc vanished");
     addDoc(first.documentId, { ...state, filePath: newPath });
 
-    const second = await openFileByPath(newPath);
+    const second = await openFromDisk(newPath);
     expect(second.alreadyOpen).toBe(true);
     expect(second.documentId).toBe(first.documentId);
     // Exactly one tab for this file — the fallback prevented a duplicate.
@@ -154,11 +150,11 @@ describe("openFileByPath — duplicate detection", () => {
   });
 });
 
-describe("openFileByPath — result structure", () => {
+describe("openFromDisk — result structure", () => {
   it("provides token and page estimates", async () => {
     const f = path.join(tmpDir, "est.md");
     await fs.writeFile(f, "Hello world, this is content.");
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
     expect(result.tokenEstimate).toBeGreaterThan(0);
     expect(result.pageEstimate).toBeGreaterThanOrEqual(1);
   });
@@ -166,17 +162,17 @@ describe("openFileByPath — result structure", () => {
   it("provides fileName from path", async () => {
     const f = path.join(tmpDir, "my-report.md");
     await fs.writeFile(f, "content");
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
     expect(result.fileName).toBe("my-report.md");
   });
 });
 
-describe("openFileFromContent — content size limit", () => {
+describe("openFromUpload — content size limit", () => {
   it("rejects content exceeding 50MB", async () => {
     const bigContent = "x".repeat(50 * 1024 * 1024 + 1);
 
     try {
-      await openFileFromContent("big.md", bigContent);
+      await openFromUpload("big.md", bigContent);
       expect.unreachable("should have thrown");
     } catch (err: unknown) {
       const e = err as NodeJS.ErrnoException;
@@ -186,27 +182,27 @@ describe("openFileFromContent — content size limit", () => {
   });
 });
 
-describe("openFileFromContent — upload properties", () => {
+describe("openFromUpload — upload properties", () => {
   it("marks uploaded files as read-only", async () => {
-    const result = await openFileFromContent("notes.md", "# Notes\nSome content");
+    const result = await openFromUpload("notes.md", "# Notes\nSome content");
     expect(result.readOnly).toBe(true);
     expect(result.source).toBe("upload");
   });
 
   it("creates synthetic upload:// path", async () => {
-    const result = await openFileFromContent("upload.txt", "text content");
+    const result = await openFromUpload("upload.txt", "text content");
     expect(result.filePath).toMatch(/^upload:\/\//);
     expect(result.filePath).toContain("upload.txt");
   });
 
   it("rejects unsupported extensions in uploads", async () => {
-    await expect(openFileFromContent("data.csv", "a,b,c")).rejects.toThrow("Unsupported");
+    await expect(openFromUpload("data.csv", "a,b,c")).rejects.toThrow("Unsupported");
   });
 });
 
-describe("openFileFromContent — Buffer content", () => {
+describe("openFromUpload — Buffer content", () => {
   it("accepts string content for text formats", async () => {
-    const result = await openFileFromContent("test.txt", "tiny content");
+    const result = await openFromUpload("test.txt", "tiny content");
     expect(result).toBeDefined();
     expect(result.format).toBe("txt");
   });
@@ -233,12 +229,12 @@ describe("SUPPORTED_EXTENSIONS", () => {
 // readOnly option
 // ---------------------------------------------------------------------------
 
-describe("openFileByPath — readOnly option", () => {
+describe("openFromDisk — readOnly option", () => {
   it("opens a file as read-only when readOnly:true is passed", async () => {
     const f = path.join(tmpDir, "ro.md");
     await fs.writeFile(f, "# Read-only content");
 
-    const result = await openFileByPath(f, { readOnly: true });
+    const result = await openFromDisk(f, { readOnly: true });
 
     expect(result.readOnly).toBe(true);
     expect(result.alreadyOpen).toBe(false);
@@ -252,7 +248,7 @@ describe("openFileByPath — readOnly option", () => {
     const f = path.join(tmpDir, "rw.md");
     await fs.writeFile(f, "# Writable content");
 
-    const result = await openFileByPath(f);
+    const result = await openFromDisk(f);
 
     expect(result.readOnly).toBe(false);
   });
@@ -262,11 +258,11 @@ describe("openFileByPath — readOnly option", () => {
     await fs.writeFile(f, "# Upgrade to read-only");
 
     // Open normally (writable)
-    const first = await openFileByPath(f);
+    const first = await openFromDisk(f);
     expect(first.readOnly).toBe(false);
 
     // Re-open with readOnly:true — should upgrade in-place
-    const second = await openFileByPath(f, { readOnly: true });
+    const second = await openFromDisk(f, { readOnly: true });
     expect(second.alreadyOpen).toBe(true);
     expect(second.readOnly).toBe(true);
 
@@ -285,10 +281,10 @@ describe("openFileByPath — readOnly option", () => {
     await fs.writeFile(f, "# No downgrade");
 
     // Open with readOnly:true
-    await openFileByPath(f, { readOnly: true });
+    await openFromDisk(f, { readOnly: true });
 
     // Re-open without specifying readOnly — should NOT downgrade
-    const second = await openFileByPath(f);
+    const second = await openFromDisk(f);
     expect(second.alreadyOpen).toBe(true);
     // readOnly reflects the re-open's derived value (false for .md), but the
     // registry and Y.Doc should NOT have been downgraded
@@ -346,7 +342,7 @@ describe("handleOpen — readOnly propagation", () => {
     await fs.writeFile(f, "# Already-open read-only upgrade");
 
     // Open the document first (writable)
-    await openFileByPath(f);
+    await openFromDisk(f);
 
     let capturedBody: unknown;
     const req = {
@@ -367,7 +363,7 @@ describe("handleOpen — readOnly propagation", () => {
   });
 });
 
-describe("openFileByPath — UNC paths are rejected BEFORE canonicalization", () => {
+describe("openFromDisk — UNC paths are rejected BEFORE canonicalization", () => {
   afterEach(() => {
     vi.restoreAllMocks();
   });
@@ -390,7 +386,7 @@ describe("openFileByPath — UNC paths are rejected BEFORE canonicalization", ()
     });
 
     try {
-      await openFileByPath(badPath);
+      await openFromDisk(badPath);
       expect.unreachable("should have thrown");
     } catch (err: unknown) {
       expect((err as NodeJS.ErrnoException).code).toBe("INVALID_PATH");
