+++ tests/server/reload-from-disk-persistence.test.ts
@@ -53,10 +53,10 @@ vi.mock("../../src/server/notifications.js", async (importOriginal) => {
   return { ...actual, pushNotification: vi.fn() };
 });
 
+import { openFromDisk } from "../../src/server/documents/open.js";
 import { removeDoc, setActiveDocId } from "../../src/server/documents/registry-testing.js";
 import { docIdFromPath } from "../../src/server/mcp/document-model.js";
 import { getOpenDocs } from "../../src/server/mcp/document-service.js";
-import { openFileByPath } from "../../src/server/mcp/file-opener.js";
 import { anchoredRange, refreshRange } from "../../src/server/positions.js";
 import { getOrCreateDocument } from "../../src/server/yjs/provider.js";
 import { Y_MAP_ANNOTATIONS } from "../../src/shared/constants.js";
@@ -91,7 +91,7 @@ async function setupOpenedFile(initialText: string): Promise<{
 }> {
   const filePath = path.join(tmpDir, "doc.md");
   await fs.writeFile(filePath, initialText, "utf-8");
-  await openFileByPath(filePath);
+  await openFromDisk(filePath);
 
   const docId = docIdFromPath(filePath);
   const doc = getOrCreateDocument(docId);
@@ -99,7 +99,7 @@ async function setupOpenedFile(initialText: string): Promise<{
   // The watcher mock captured the (filePath, callback) — driving the callback
   // is exactly what fs.watch does on a real on-disk change.
   const lastCall = watcherMocks.watchFile.mock.calls.at(-1);
-  if (!lastCall) throw new Error("watchFile was not called by openFileByPath");
+  if (!lastCall) throw new Error("watchFile was not called by openFromDisk");
   const onChanged = lastCall[1] as (p: string) => Promise<void>;
   const triggerReload = async () => {
     await onChanged(filePath);
